requirements.txt
beautifulsoup4==4.7.1
Office365-REST-Python-Client==2.1.1
